Class that extends ShinyDriver from

Class that extends ShinyDriver from shinytest package for load testing.

ShinyLoadDriver

Format

An object of class R6ClassGenerator of length 24.

Usage

app <- ShinyLoadDriver$new(url, connectionId) app$snapshotInit("myloadtest") app$setInputs(...) app$snapshot() app$stop() app$getEventLog()

Arguments

path

The url of a deployed Shiny app starting with 'http://' or https://'

connectionId

Identifies the current connection, typically a sequential id

...

ShinyLoadDriver inherits the other arguments from ShinyDriver

Details

app$new() creates a ShinyLoadDriver object. It opens a session in phantomJS connected to the Shiny app deployed at path. Because ShinyLoadDriver connects to a deployed application, not all of the methods from ShinyDriver are available.

app$setInputs() sets an input value in the Shiny app and records the amount of time before either an output is updated or a timeout occurs. The timeout value, in milliseconds, can be specified using setInputs(..., timeout_ = 3000). setInputs() accepts at most 1 input at a time.

app$snapshot() takes a screenshot of the application and saves it to a sub-directory of the current working directory. The subdirectory is named based on app$snapshotInit("directoryName"). Screenshots are named connectionId_snapshotCount.png where snapshotCount is a counter incrementing with each screenshot taken by the ShinyLoadDriver.

app$getEventLog() returns an event log with information on the test.