Record a Session for Load Test

Record a Session for Load Test

record_session(target_app_url, host = "127.0.0.1", port = 8600,
  output_file = "recording.log", open_browser = TRUE)

Arguments

target_app_url
The URL of the deployed application

host
The host where the proxy will run. Usually localhost is used.

port
The port for the proxy. Default is 8600. Change this default if port 8600 is used by another service.

output_file
The name for the generated log file.

open_browser
Whether to open a browser on the proxy (default=TRUE) or not (FALSE).

Value

Creates a recording file that can be used to drive a load test.

Details

This function creates a reverse proxy listening at host:port that stands in front of target_app_url. As the user interacts with the app, the actions and responses are recorded in a file that can be replayed later.